Fix max-classes-per-file violations by splitting up modules
