+++ packages/table/src/interactions/menus/menuContext.tsx
@@ -5,8 +5,9 @@
 
 import { IMenuItemProps, MenuItem } from "@blueprintjs/core";
 import * as React from "react";
-import { Clipboard } from "../common/clipboard";
-import { ICellCoordinate, IRegion, Regions } from "../regions";
+
+import { Clipboard } from "../../common/clipboard";
+import { ICellCoordinate, IRegion, Regions } from "../../regions";
 
 export type IContextMenuRenderer = (context: IMenuContext) => JSX.Element;
 
@@ -65,46 +66,5 @@ export class MenuContext implements IMenuContext {
 
     public getUniqueCells() {
         return Regions.enumerateUniqueCells(this.regions, this.numRows, this.numCols);
-    }
-}
-
-export interface ICopyCellsMenuItemProps extends IMenuItemProps {
-    /**
-     * The `IMenuContext` that launched the menu.
-     */
-    context: IMenuContext;
-
-    /**
-     * A callback that returns the data for a specific cell. This need not
-     * match the value displayed in the `<Cell>` component. The value will be
-     * invisibly added as `textContent` into the DOM before copying.
-     */
-    getCellData: (row: number, col: number) => any;
-
-    /**
-     * If you want to do something after the copy or if you want to notify the
-     * user if a copy fails, you may provide this optional callback.
-     *
-     * Due to browser limitations, the copy can fail. This usually occurs if
-     * the selection is too large, like 20,000+ cells. The copy will also fail
-     * if the browser does not support the copy method (see
-     * `Clipboard.isCopySupported`).
-     */
-    onCopy?: (success: boolean) => void;
-}
-
-export class CopyCellsMenuItem extends React.Component<ICopyCellsMenuItemProps, {}> {
-    public render() {
-        return <MenuItem {...this.props} onClick={this.handleClick} />;
-    }
-
-    private handleClick = () => {
-        const { context, getCellData, onCopy } = this.props;
-        const cells = context.getUniqueCells();
-        const sparse = Regions.sparseMapCells(cells, getCellData);
-        const success = Clipboard.copyCells(sparse);
-        if (onCopy != null) {
-            onCopy(success);
-        }
     }
 }
